==> 2-pre-process-images.ipynb <==
import os, sys
from PIL import Image
%store -r IMAGE_SIZE

path="dataset/"
output_path="dataset-pre-process/"
dirs = os.listdir(path)
i = 0
for sub_dir in dirs:
    for item in os.listdir(path + sub_dir):
        if os.path.isfile(path + sub_dir + '/' + item):
            im = Image.open(path + sub_dir + '/' + item)
            f, e = os.path.splitext(path + sub_dir + '/' + item)
            imResize = im.resize((IMAGE_SIZE,IMAGE_SIZE), Image.ANTIALIAS)
            rgb_im = imResize.convert('RGB')
            if not os.path.exists(output_path):
                os.mkdir(output_path)
            if not os.path.exists(output_path + sub_dir):
                os.mkdir(output_path + sub_dir) 
            rgb_im.save(output_path + sub_dir + '/' + str(i) + '.jpg', 'JPEG',optimize=True, quality=20)
            i = i + 1
