==> 4-data-augmentation.ipynb <==
import Augmentor as aug
import glob
import os
import numpy as np
import cv2 as cv
import PIL
from Augmentor.Operations import Operation
%store -r IMAGE_SIZE


class Lightning(Operation):
    def __init__(self, probability, intensity_low=0.7, intensity_high=1.2):
        Operation.__init__(self, probability)
        self.intensity_low = intensity_low
        self.intensity_high = intensity_high

    def perform_operation(self, images):
        for i, image in enumerate(images):
            image = np.array(image.convert('RGB'))
            row, col, _ = image.shape
            light_intensity = np.random.randint(
                int(self.intensity_low * 100),
                int(self.intensity_high * 100)
            )
            light_intensity /= 100

            gaussian = 100 * np.random.random((row, col, 1))
            gaussian = np.array(gaussian, dtype=np.uint8)
            gaussian = np.concatenate((gaussian, gaussian, gaussian), axis=2)

            image = cv.addWeighted(image, light_intensity, gaussian, 0.25, 0)

            image = PIL.Image.fromarray(image)
            images[i] = image
            
        return images

DIRECTORIES = 'dataset-split/*/*'
folders = []
for f in glob.glob(DIRECTORIES):
    print(f[14:16])
    if os.path.isdir(f):
        if(f[14:16] == "va"):
            folders.append(os.path.abspath(f)+"_va")
        elif(f[14:16] == "tr"):
            folders.append(os.path.abspath(f)+"_tr")
        elif(f[14:16] == "te"):
            folders.append(os.path.abspath(f)+"_te")

print('Classes found {}'.format([os.path.split(x)[1] for x in folders]))


pipelines = {}
for folder in folders:
    print(folder[:-2])
    pipelines[os.path.split(folder)[1]] = \
        (aug.Pipeline(source_directory=folder[:-3],
            output_directory='.',
            save_format='jpg'
            )
        )

classes_count = []
for p in pipelines.values():
    print("Class '{}' has {} samples".format(
        p.augmentor_images[0].class_label,
        len(p.augmentor_images)
    ))
    classes_count.append(len(p.augmentor_images))

print(classes_count)

lightning = Lightning(probability=0.3)

for p in pipelines.values():
    p.rotate(probability=0.5, max_left_rotation=10, max_right_rotation=10)
    # p.zoom_random(probability=0.4, percentage_area=0.9)
    p.flip_random(probability=0.5)
    p.flip_left_right(probability=0.7)
    p.flip_top_bottom(probability=0.5)
    p.shear(probability=0.5, max_shear_left=10, max_shear_right=10)
    p.random_distortion(probability=0.8, grid_width=5, grid_height=5, magnitude=15)
    p.add_operation(lightning)
    p.resize(probability=1.0, width=IMAGE_SIZE, height=IMAGE_SIZE)

SAME_SAMPLING = False
MULTIPLIER = 29

for p in pipelines.values():
    if SAME_SAMPLING:
        diff = max(classes_count) - len(p.augmentor_images)
        p.sample((len(p.augmentor_images) + diff)*MULTIPLIER + diff)
    else:
        p.sample(len(p.augmentor_images)*MULTIPLIER)
