==> 5-create_label_tensorflow.ipynb <==
import os
import glob
import pandas as pd

lesions = {"Basal Cell Carcinoma": 0, 
    "Lentigo": 1,
    "Malignant Melanoma": 2, 
    "Melanocytic naevus": 3,
    "seborrhoeic keratosis": 4,
    "Wart": 5, 
    "Actinic Keratosis": 6,
    "Squamous Cell Carcinoma": 7,
    "Intraepithelial Carcinoma": 8, 
    "Pyogenic Granuloma": 9,
    "Haemangioma": 10,
    "Dermatofibroma": 11
}

folder_test="dataset-split/test/*/*.jpg"
folder_train="dataset-split/train/*/*.jpg"
folder_val="dataset-split/val/*/*.jpg"


def train_test_val_csv(folder):
    resnet_augmented = folder[:-5] + "/*/*.jpg"

    lession_list = []
    img_list = []

    for filename in glob.glob(folder):
        folder, img = os.path.split(filename)
        _, lession = os.path.split(folder)

        lession_list.append(lesions[lession])
        img_list.append(img[:-4])

    for filename in glob.glob(resnet_augmented):
        folder, img = os.path.split(filename)
        folder, _ = os.path.split(folder)
        _, lession = os.path.split(folder)

        lession_list.append(lesions[lession])
        img_list.append(img[:-4])

    data = {'id': img_list, 'label':lession_list}
    
    return pd.DataFrame(data=data)

train_test_val_csv(folder_test).to_csv(path_or_buf="./test.csv")
train_test_val_csv(folder_train).to_csv(path_or_buf="./train.csv")
train_test_val_csv(folder_val).to_csv(path_or_buf="./val.csv")
